3813. Vowel-Consonant Score

+++ Problems/3013. Divide an Array Into Subarrays With Minimum Cost II/3013_divide_an_array_into_subarrays_with_minimum_cost_ii_cpp/main.cpp
@@ -75,7 +75,7 @@ void printArray(vector<int> array) {
     cout << endl;
 }
 
-void test(vector<int> nums, int k, int dist, int expected) {
+void test(vector<int> nums, int k, int dist, long long expected) {
     cout << "Array: ";
     printArray(nums);
